Split unitests into modules
